backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
python-dotenv==1.0.0
marshmallow==3.20.1
google-generativeai==0.3.2
APScheduler==3.10.4